requirements.txt
streamlit
opencv-python
mediapipe
pyttsx3
google-generativeai
numpy
